Rename consoleColor to consoleColours
